Добавил и подключил статическую библитеку к проектам, убрал дублирующиеся файлы из проектов
