+++ labs/03/samples_win32/libgl/framework.h
@@ -1,0 +1,12 @@
+﻿#pragma once
+
+#define WIN32_LEAN_AND_MEAN             // Исключите редко используемые компоненты из заголовков Windows
+
+// Файлы заголовков Windows
+#include <windows.h>
+// Файлы заголовков среды выполнения C
+//#include <stdlib.h>
+//#include <malloc.h>
+//#include <memory.h>
+//#include <tchar.h>
+#include <gl\gl.h>
