+++ labs/03/samples_win32/libgl/pch.cpp
@@ -1,0 +1,5 @@
+﻿// pch.cpp: файл исходного кода, соответствующий предварительно скомпилированному заголовочному файлу
+
+#include "pch.h"
+
+// При использовании предварительно скомпилированных заголовочных файлов необходим следующий файл исходного кода для выполнения сборки.
