+++ labs/03/samples_win32/samples_winapi.sln
@@ -27,6 +27,8 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "10.polygon", "10.polygon\10
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "11.rotation", "11.rotation\11.rotation.vcxproj", "{9C19D266-88B3-4FE3-8EEC-AA73F4E8A76A}"
 EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "libgl", "libgl\libgl.vcxproj", "{4E843E75-EC1E-4423-ABC1-6AC5069704DA}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x64 = Debug|x64
@@ -131,6 +133,14 @@ Global
 		{9C19D266-88B3-4FE3-8EEC-AA73F4E8A76A}.Release|x64.Build.0 = Release|x64
 		{9C19D266-88B3-4FE3-8EEC-AA73F4E8A76A}.Release|x86.ActiveCfg = Release|Win32
 		{9C19D266-88B3-4FE3-8EEC-AA73F4E8A76A}.Release|x86.Build.0 = Release|Win32
+		{4E843E75-EC1E-4423-ABC1-6AC5069704DA}.Debug|x64.ActiveCfg = Debug|x64
+		{4E843E75-EC1E-4423-ABC1-6AC5069704DA}.Debug|x64.Build.0 = Debug|x64
+		{4E843E75-EC1E-4423-ABC1-6AC5069704DA}.Debug|x86.ActiveCfg = Debug|Win32
+		{4E843E75-EC1E-4423-ABC1-6AC5069704DA}.Debug|x86.Build.0 = Debug|Win32
+		{4E843E75-EC1E-4423-ABC1-6AC5069704DA}.Release|x64.ActiveCfg = Release|x64
+		{4E843E75-EC1E-4423-ABC1-6AC5069704DA}.Release|x64.Build.0 = Release|x64
+		{4E843E75-EC1E-4423-ABC1-6AC5069704DA}.Release|x86.ActiveCfg = Release|Win32
+		{4E843E75-EC1E-4423-ABC1-6AC5069704DA}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
